requirements.txt
aniso8601==9.0.1
APScheduler==3.7.0
asttokens==2.0.5
certifi==2020.12.5
chardet==4.0.0
click 
colorama==0.4.4
dnspython==1.16.0
eventlet==0.30.2
executing==0.6.0
Flask==1.1.2
Flask-JWT-Extended==4.2.0
Flask-RESTful==0.3.8
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.1.0
icecream==2.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
luxand==0.7
MarkupSafe==2.0.1
numpy==1.20.2
opencv-python==4.5.1.48
pandas==1.2.5
passlib==1.7.4
psycogreen==1.0.2
psycopg2-binary 
Pygments==2.8.1
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
Shapely==1.7.1
six==1.15.0
tzlocal==2.1
urllib3==1.26.4
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
